Create CodonCount
This code uses a HashMap to find the break a strand of DNA, read in from a file, into consecutive codons, based on the reading frame designated (start index of 0,1,or 2). It then tells you which codon is most common and prints out all of the codons found between two designated numbers.